# openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User","description":"User management APIs"}],"paths":{"/users/{id}":{"get":{"tags":["User"],"summary":"Get a user by ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["User"],"summary":"Update a user","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"User updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/users":{"post":{"tags":["User"],"summary":"Create a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"User created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/companysettings":{"post":{"tags":["Company Settings"],"summary":"Save company settings","operationId":"saveSettings","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanySettings"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/companysettings/unit-id/{unitId}":{"get":{"tags":["Company Settings"],"summary":"Get settings by unit ID","operationId":"getByUnitId","parameters":[{"name":"unitId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanySettings"}}}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"}}},"CompanySettings":{"type":"object","properties":{"unitId":{"type":"string"},"settingName":{"type":"string"},"settingValue":{"type":"string"},"displayName":{"type":"string"}}}}}}